SUNWAY: fix bugs in decreasing frequency
Mainline: KYLIN-only
Severity: Low

Task: #382309

The issue arose because the downshift_freq() was attempting to scale down
the frequency of a physicial core when only one of its logical core was
offline. The correct behavior should be to scale down the frequency only
when both logical cores of a pyhsicial core are offline.

K2CI-Arch: Sunway
Signed-off-by: Tang Jinyang <tangjinyang@wxiat.com>
Signed-off-by: liuyunchen <liuyunchen@kylinos.cn>
Change-Id: I5e7ed9071f8daca3f78a0ee1ae3455020d2ae26c
Reviewed-on: https://code.kylinos.cn:443/c/institute/kernel/klinux/+/66195
Reviewed-by: 刘云 <liuyun01@kylinos.cn>
Code-Style: kernel-bot <kernel-bot@kylinos.cn>
Tested-by: kernel-bot <kernel-bot@kylinos.cn>