arm64: acpi: Harden get_cpu_for_acpi_id() against missing CPU entry
Mainline: 2488444274c70038eb6b686cba5f1ce48ebb9cdd
From: v6.10-rc3-12-g2488444274c7
Severity: Moderate

task: 327586

In a review discussion of the changes to support vCPU hotplug where
a check was added on the GICC being enabled if was online, it was
noted that there is need to map back to the cpu and use that to index
into a cpumask. As such, a valid ID is needed.

If an MPIDR check fails in acpi_map_gic_cpu_interface() it is possible
for the entry in cpu_madt_gicc[cpu] == NULL.  This function would
then cause a NULL pointer dereference.   Whilst a path to trigger
this has not been established, harden this caller against the
possibility.

Reviewed-by: Gavin Shan <gshan@redhat.com>
Signed-off-by: Jonathan Cameron <Jonathan.Cameron@huawei.com>
Link: https://lore.kernel.org/r/20240529133446.28446-13-Jonathan.Cameron@huawei.com
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>

(backported from commit 2488444274c70038eb6b686cba5f1ce48ebb9cdd)

Signed-off-by: Kunwu Chan <chentao@kylinos.cn>
Change-Id: I3ee694bd053d84f13747d84c5988f6dabdac5c1b
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/123515
Reviewed-by: Jackie Liu <liuyun01@kylinos.cn>
Tested-by: k2ci <kernel-bot@kylinos.cn>
Code-Style: k2ci <kernel-bot@kylinos.cn>