KYLIN: net/hnic3: Modify the initial value of the g_uld_info struct array
Mainline: KYLIN-only
Severity: Low

Task: 222454

When compiling the kernel with 'make allyesconfig', the compiler
sets '-Werror= designing-init'. At this time, if the member
variables of the structure are all function Pointers, the member
variables of the structure need to be initialized with the form
'.member_name = value'. Or directly do not initialize.that can
solve the compiler error problem

K2CI-Arch: All
Signed-off-by: taowei <taowei@kylinos.cn>
Change-Id: Id7c32cf3ea0f38edea4dabc5e26d195325a3cbba
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/114742
Reviewed-by: Jackie Liu <liuyun01@kylinos.cn>
Code-Style: k2ci <kernel-bot@kylinos.cn>
Tested-by: k2ci <kernel-bot@kylinos.cn>