KYLIN: vcpu_stall_detector: add vcpu stall detector
Mainline: KYLIN-only
Severity: Low

task: #218469

This device driver acts as a soft lockup detector by relying on
the host backend driver to measure the elapesed time between
subsequent 'pet' events. If the elapsed time doesn't match an
expected value, the backend driver decides that the guest vCPU
is locked and resets the guest. The communication with the backend
driver which is mounted on the pci bus is done through VIRTIO.

Each vcpu starts a monitoring thread that periodically sends
dog feeding information to the qemu.

Signed-off-by: Zhang Hao <zhanghao1@kylinos.cn>
Change-Id: I0bc98ab931b82ac0dd20e429a2e9a44d513c7cf3
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/116401
Code-Style: k2ci <kernel-bot@kylinos.cn>
Tested-by: k2ci <kernel-bot@kylinos.cn>
Reviewed-by: Jackie Liu <liuyun01@kylinos.cn>