wifi: cfg80211: wext: set ssids=NULL for passive scans
Mainline: 0941772342d59e48733131ac3a202fa1a4d832e9
From: v6.10-rc5
Severity: Moderate

task: 337947

commit 0941772342d59e48733131ac3a202fa1a4d832e9 upstream.

In nl80211, we always set the ssids of a scan request to
NULL when n_ssids==0 (passive scan). Drivers have relied
on this behaviour in the past, so we fixed it in 6 GHz
scan requests as well, and added a warning so we'd have
assurance the API would always be called that way.

syzbot found that wext doesn't ensure that, so we reach
the check and trigger the warning. Fix the wext code to
set the ssids pointer to NULL when there are none.

Reported-by: syzbot+cd6135193ba6bb9ad158@syzkaller.appspotmail.com
Fixes: f7a8b10bfd61 ("wifi: cfg80211: fix 6 GHz scan request building")
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>

(backported from commit de3994a44ebc6162218dcbc2eeb850dee2b9c277)

K2CI-Arch: None
Signed-off-by: k2ci <kernel-bot@kylinos.cn>
Change-Id: I363c10f4b4c3256aea98b4fb85d706dafabe8b3d
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/125088
Reviewed-by: Jackie Liu <liuyun01@kylinos.cn>